Non-parametric tests for the two-sample problem based on order statistics and power comparisons

tnl.test performs a nonparametric test for two sample test on vectors of data.

ptnl gives the distribution function of \(T_n^{(\ell)}\) against the specified quantiles.

dtnl gives the density of \(T_n^{(\ell)}\) against the specified quantiles.

qtnl gives the quantile function of \(T_n^{(\ell)}\) against the specified probabilities.

rtnl generates random values from \(T_n^{(\ell)}\).

dtnl.lehmann gives the density of \(T_n^{(\ell)}\) under Lehmann alternatives.

ptnl.lehmann gives the distribution function of \(T_n^{(\ell)}\) under Lehmann alternatives.

tnl_mean gives an expression for \(E(T_n^{(\ell)})\) under \(H_0: F = G\).

tnl.test(x, y, l, exact = "NULL")

ptnl(q, n, l, exact = "NULL", trial = 1e+05)

dtnl(k, n, l, exact = "NULL", trial = 1e+05)

qtnl(p, n, l, exact = "NULL", trial = 1e+05)

rtnl(N, n, l)

dtnl.lehmann(l, n, gamma)

ptnl.lehmann(l, n, gamma)

tnl_mean(n, l)

Arguments

x
the first (non-empty) numeric vector of data values.

y
the second (non-empty) numeric vector of data values.

l
class parameter of \(T_n^{(\ell)}\)

exact
the method that will be used. "NULL" or a logical indicating whether an exact should be computed. See ‘Details’ for the meaning of NULL.

n
sample size.

trial
number of trials for simulation.

k, q
vector of quantiles.

p
vector of probabilities.

N
number of observations. If length(N) > 1, the length is taken to be the number required.

gamma
parameter of Lehmann alternative

Value

tnl.test returns a list with the following components

statistic:

the value of the test statistic.

p.value:

the p-value of the test.

ptnl returns a list with the following components

method:

The method that was used (exact or simulation). See ‘Details’.

cdf:

distribution function of \(T_n^{(\ell)}\) against the specified quantiles.

dtnl returns a list with the following components

method:

The method that was used (exact or simulation). See ‘Details’.

pmf:

density of \(T_n^{(\ell)}\) against the specified quantiles.

qtnl returns a list with the following components

method:

The method that was used (exact or simulation). See ‘Details’.

quantile:

quantile function against the specified probabilities.

rtnl return N of the generated random values.

dtnl.lehmann return vector of the density under Lehmann alternatives against the specified gamma.

ptnl.lehmann return vector of the distribution under Lehmann alternatives against the specified gamma.

tnl_mean return the mean of \(T_n^{(\ell)}\).

Details

A non-parametric two-sample test is performed for testing null hypothesis \(H_0: F = G\) against the alternative hypothesis \(H_1:F \not= G\). The assumptions of the \(T_n^{(\ell)}\) test are that both samples should come from a continuous distribution and the samples should have the same sample size.
Missing values are silently omitted from x and y.
Exact and simulated p-values are available for the \(T_n^{(\ell)}\) test. If exact ="NULL" (the default) the p-value is computed based on exact distribution when the sample size is less than 11. Otherwise, p-value is computed based on a Monte Carlo simulation. If exact ="TRUE", an exact p-value is computed. If exact="FALSE" , a Monte Carlo simulation is performed to compute the p-value. It is recommended to calculate the p-value by a Monte Carlo simulation (use exact="FALSE"), as it takes too long to calculate the exact p-value when the sample size is greater than 10.
The probability mass function (pmf), cumulative density function (cdf) and quantile function of \(T_n^{(\ell)}\) are also available in this package, and the above-mentioned conditions about exact ="NULL", exact ="TRUE" and exact="FALSE" is also valid for these functions.
Exact distribution of \(T_n^{(\ell)}\) test is also computed under Lehman alternative.
Random number generator of \(T_n^{(\ell)}\) test statistic are provided under null hypothesis in the library.

References

Karakaya K. et al. (2021). A Class of Non-parametric Tests for the Two-Sample Problem based on Order Statistics and Power Comparisons . Submitted paper.
Aliev F. et al. (2021). A Nonparametric Test for the Two-Sample Problem based on Order Statistics. Submitted paper.

Examples

require(stats)
x <- rnorm(7, 2, 0.5)
y <- rnorm(7, 0, 1)
tnl.test(x, y, l = 2)
#> $statistic
#> [1] 2
#> 
#> $p.value
#> [1] 0.02447552
#> 
# $statistic
# [1] 2
#
# $p.value
# [1] 0.02447552
##############
ptnl(q = 2, n = 6, l = 2, trial = 100000)
#> $method
#> [1] "exact"
#> 
#> $cdf
#> [1] 0.03030303
#> 
# $method
# [1] "exact"
#
# $cdf
# [1] 0.03030303
##############
dtnl(k = 3, n = 7, l = 2, trial = 100000)
#> $method
#> [1] "exact"
#> 
#> $pmf
#> [1] 0.05710956
#> 
# $method
# [1] "exact"
#
# $pmf
# [1] 0.05710956
if (FALSE) {
qtnl(.3, 4, 1, exact = "FALSE")
# $method
# [1] "Monte Carlo simulation"
#
# $quantile
# [1] 2
}
##############
rtnl(N = 15, n = 7, l = 2)
#>  [1] 7 5 6 6 7 3 6 6 7 6 6 4 7 3 7
# [1] 6 7 6 6 3 5 7 7 7 7 7 6 5 7 4
##############
dtnl.lehmann(l = 2, n = 6, gamma = 0.8)
#> [1] 0.00000000 0.03757203 0.08230829 0.14229514 0.23276690 0.50505764
# [1] 0.00000000 0.03757203 0.08230829 0.14229514 0.23276690 0.50505764
##############
ptnl.lehmann(l = 2, 5, gamma = 1.2)
#> [1] 0.0000000 0.0444164 0.1529147 0.3694389 1.0000000
# [1] 0.0000000 0.0444164 0.1529147 0.3694389 1.0000000
##############
require(base)
tnl_mean(11, 2)
#> [1] 8.058115
# [1] 8.058115